fix: get connect data when verify modal is opened
