--- src/modules/Verification/containers/KYC/index.types.ts
@@ -1,6 +1,0 @@
-import { KYCStatus } from 'src/core/adaptors';
-
-export interface KYCProps {
-  connectUrl: string;
-  status: KYCStatus;
-}
